Moved CreateInputLayout helper to DirectXHelpers renamed as CreateInputLayoutFromEffect
