+++ Inc/DirectXHelpers.h
@@ -54,15 +54,17 @@
 
 namespace DirectX
 {
+    class IEffect;
+
     // simliar to std::lock_guard for exception-safe Direct3D resource locking
     class MapGuard : public D3D11_MAPPED_SUBRESOURCE
     {
     public:
         MapGuard(_In_ ID3D11DeviceContext* context,
-                 _In_ ID3D11Resource *resource,
-                 _In_ UINT subresource,
-                 _In_ D3D11_MAP mapType,
-                 _In_ UINT mapFlags) noexcept(false)
+            _In_ ID3D11Resource *resource,
+            _In_ UINT subresource,
+            _In_ D3D11_MAP mapType,
+            _In_ UINT mapFlags) noexcept(false)
             : mContext(context), mResource(resource), mSubresource(subresource)
         {
             HRESULT hr = mContext->Map(resource, subresource, mapType, mapFlags, this);
@@ -110,43 +112,43 @@ namespace DirectX
 
     // Helper sets a D3D resource name string (used by PIX and debug layer leak reporting).
     template<UINT TNameLength>
-    inline void SetDebugObjectName(_In_ ID3D11DeviceChild* resource, _In_z_ const char (&name)[TNameLength]) noexcept
+    inline void SetDebugObjectName(_In_ ID3D11DeviceChild* resource, _In_z_ const char(&name)[TNameLength]) noexcept
     {
-        #if !defined(NO_D3D11_DEBUG_NAME) && ( defined(_DEBUG) || defined(PROFILE) )
-            #if defined(_XBOX_ONE) && defined(_TITLE)
-                wchar_t wname[MAX_PATH];
-                int result = MultiByteToWideChar(CP_UTF8, 0, name, TNameLength, wname, MAX_PATH);
-                if (result > 0)
-                {
-                    resource->SetName(wname);
-                }
-            #else
-                resource->SetPrivateData(WKPDID_D3DDebugObjectName, TNameLength - 1, name);
-            #endif
-        #else
-            UNREFERENCED_PARAMETER(resource);
-            UNREFERENCED_PARAMETER(name);
-        #endif
+#if !defined(NO_D3D11_DEBUG_NAME) && ( defined(_DEBUG) || defined(PROFILE) )
+#if defined(_XBOX_ONE) && defined(_TITLE)
+        wchar_t wname[MAX_PATH];
+        int result = MultiByteToWideChar(CP_UTF8, 0, name, TNameLength, wname, MAX_PATH);
+        if (result > 0)
+        {
+            resource->SetName(wname);
+        }
+#else
+        resource->SetPrivateData(WKPDID_D3DDebugObjectName, TNameLength - 1, name);
+#endif
+#else
+        UNREFERENCED_PARAMETER(resource);
+        UNREFERENCED_PARAMETER(name);
+#endif
     }
 
     template<UINT TNameLength>
-    inline void SetDebugObjectName(_In_ ID3D11DeviceChild* resource, _In_z_ const wchar_t (&name)[TNameLength])
+    inline void SetDebugObjectName(_In_ ID3D11DeviceChild* resource, _In_z_ const wchar_t(&name)[TNameLength])
     {
-        #if !defined(NO_D3D11_DEBUG_NAME) && ( defined(_DEBUG) || defined(PROFILE) )
-            #if defined(_XBOX_ONE) && defined(_TITLE)
-                resource->SetName( name );
-            #else
-                char aname[MAX_PATH];
-                int result = WideCharToMultiByte(CP_UTF8, 0, name, TNameLength, aname, MAX_PATH, nullptr, nullptr);
-                if (result > 0)
-                {
-                    resource->SetPrivateData(WKPDID_D3DDebugObjectName, TNameLength - 1, aname);
-                }
-            #endif
-        #else
-            UNREFERENCED_PARAMETER(resource);
-            UNREFERENCED_PARAMETER(name);
-        #endif
+#if !defined(NO_D3D11_DEBUG_NAME) && ( defined(_DEBUG) || defined(PROFILE) )
+#if defined(_XBOX_ONE) && defined(_TITLE)
+        resource->SetName(name);
+#else
+        char aname[MAX_PATH];
+        int result = WideCharToMultiByte(CP_UTF8, 0, name, TNameLength, aname, MAX_PATH, nullptr, nullptr);
+        if (result > 0)
+        {
+            resource->SetPrivateData(WKPDID_D3DDebugObjectName, TNameLength - 1, aname);
+        }
+#endif
+#else
+        UNREFERENCED_PARAMETER(resource);
+        UNREFERENCED_PARAMETER(name);
+#endif
     }
 
     // Helper to check for power-of-2
@@ -176,5 +178,20 @@ namespace DirectX
             return (size + mask) & ~mask;
         }
         return size;
+    }
+
+    // Helper for creating a Direct3D input layout to match a shader from an IEffect
+    HRESULT __cdecl CreateInputLayoutFromEffect(_In_ ID3D11Device* device,
+        _In_ IEffect* effect,
+        _In_reads_(count) const D3D11_INPUT_ELEMENT_DESC* desc,
+        size_t count,
+        _Outptr_ ID3D11InputLayout** pInputLayout) noexcept;
+
+    template<typename T>
+    HRESULT CreateInputLayoutFromEffect(_In_ ID3D11Device* device,
+        _In_ IEffect* effect,
+        _Outptr_ ID3D11InputLayout** pInputLayout) noexcept
+    {
+        return CreateInputLayoutFromEffect(device, effect, T::InputElements, T::InputElementCount, pInputLayout);
     }
 }
